[#49] add post page
