+++ src/locales/us/translation.json
@@ -56,6 +56,7 @@
     },
     "topnav": {
       "main-page": "Community",
+      "post": "Post",
       "debate": "Discussion",
       "summary": "Summary",
       "search": "Explore",
